WIP - restructure for pypi packaging
