+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.0.0] 2022-10-19
+
+### Added
+
+### Changed
+  - Restructured code base to support packaging for pypi.
+  - Migrated README documentation to readthedocs format.
+
+### Removed
+
 ## [2.1.0] 2021-09-16
 ### Added
   - Support `#channel@guild_id` representation.
