--- discord.plug
@@ -1,7 +1,0 @@
-[Core]
-Name = Discord
-Module = discordb
-
-[Documentation]
-Description = This is a backend implementation for discord.
-
