--- requirements.txt
@@ -1,1 +1,0 @@
-discord.py==1.7.3
