+++ src/err-backend-discord/err-backend-discord.plug
@@ -1,0 +1,7 @@
+[Core]
+Name = Discord
+Module = err-backend-discord
+
+[Documentation]
+Description = Discord backend for Errbot.
+
